requirements.txt
st-supabase-connection
plotly